requirements.txt
flask==3.1.0
flask-cors==5.0.1
requests==2.32.3
beautifulsoup4==4.13.3
openai>=1.30.0
python-dotenv==1.1.0
